Maze Runner
This program creates a maze.  The distance of each grid cell from the
start of the maze is then marked and the shortest route from the stark
of the maze to the end of the maze is highlighted.  This is accomplished
by using a stack, and queue, and a linked list.